delete endpoint fixed

+++ webserver/endpoints/wiki-article.endpoint.js
@@ -16,7 +16,7 @@ export class WikiArticleEndpointService {
 
 
             const queryAllArticles = () => {
-                return genericQueryExecutor('SELECT * FROM articles');
+                return genericQueryExecutor('SELECT * FROM articles ORDER BY _id DESC');
             }
 
             const getArticleByUID = () => {
@@ -38,8 +38,9 @@ export class WikiArticleEndpointService {
 
             const deleteSelectedArticle = (requestData) => {
                 return new Promise(resolve => {
+                    console.log(`${requestData.credentials.username} requested a deletion of article ${requestData.articleID}`);
                     genericQueryExecutor(`SELECT * FROM users WHERE username='${requestData.credentials.username}' AND password='${requestData.credentials.password}'`).then(authResponse => {
-                        if (authResponse.queryValidation === 'valid') {
+                        if (authResponse.queryValidation === 'valid' && authResponse.values.length === 1) {
                             genericQueryExecutor(`DELETE FROM articles WHERE article_id='${requestData.articleID}' AND author='${requestData.credentials.UID}'`).then(dbResponse => {
                                 if (dbResponse.queryValidation === 'valid') resolve({ queryValidation: 'valid' });
                                 else resolve({ queryValidation: 'invalid' });
@@ -188,18 +189,6 @@ export class WikiArticleEndpointService {
 
                 case 'delete-article':
                     let deleteRequestData = req.body.values;
-                    /* 
-                    TODO: register in documentation,
-                    data structure is:
-                    {
-                        articleID : 1,
-                        credentials : {
-                            username : 'blin',
-                            password : 'blin',
-                            UID: 1
-                        }
-                    }
-                    */
                     deleteSelectedArticle(deleteRequestData).then(dbResponse => {
                         if (dbResponse.queryValidation === 'valid') {
                             const VALID_RESPONSE = {
